requirements.txt
# ============================================
# 의약품 정보 RAG 시스템 의존성
# ============================================

# --- Core LangChain ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# --- LangSmith (Tracing & Monitoring) ---
langsmith>=0.2.0

# --- Database ---
supabase>=2.0.0

# --- Web Framework ---
streamlit>=1.40.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0

# --- Data Processing ---
pandas>=2.0.0

# --- Benchmark & Evaluation ---
ragas>=0.1.0
datasets>=2.14.0
rank-bm25>=0.2.2
plotly>=5.18.0
